requirements.txt
flask
psycopg2-binary
djangorestframework-jwt
